Amazon Bedrock/RAG Project/main.py
Here are comments for the provided code, explaining each section and function's role in the Retrieval-Augmented Generation (RAG) application.

-----

### Code Comments for RAG Application

```python
import os  # Used to access environment variables.
import boto3  # The AWS SDK for Python, used to interact with AWS services like Bedrock.
import streamlit as st  # The framework for building the web-based user interface.
from langchain.llms.bedrock import Bedrock  # A LangChain wrapper for Bedrock's large language models.
from langchain.embeddings import BedrockEmbeddings  # A LangChain wrapper for Bedrock's embedding models.
from langchain.document_loaders import PyPDFDirectoryLoader  # A utility to load PDF documents from a directory.
from langchain.text_splitter import RecursiveCharacterTextSplitter  # A method to split documents into smaller chunks.
from langchain.vectorstores import Chroma  # A vector store library used to store and search document embeddings.
from langchain.prompts import PromptTemplate  # A class to create structured prompts for LLMs.
from langchain.chains import RetrievalQA  # A LangChain chain that combines document retrieval with question answering.
from dotenv import load_dotenv  # A library to load environment variables from a .env file.

# Load environment variables from the .env file. This is a best practice
# for managing sensitive information like API keys and credentials.
load_dotenv()

# Retrieve AWS credentials and region from environment variables.
aws_access_key_id = os.getenv("aws_access_key_id")
aws_secret_access_key = os.getenv("aws_secret_access_key")
region_name = os.getenv("region_name")


# Define a prompt template that provides instructions and context to the LLM.
# The template includes placeholders for the retrieved context and the user's question.
prompt_template = """

Human: Use the following pieces of context to provide a 
concise answer to the question at the end but use atleast summarize with 
250 words with detailed explantions. If you don't know the answer, 
just say that you don't know, don't try to make up an answer.
<context>
{context}
</context

Question: {question}

Assistant:"""


# Initialize a boto3 client for the Bedrock runtime service.
# The credentials and region are passed from the environment variables for secure access.
bedrock = boto3.client(
    service_name = "bedrock-runtime", 
    region_name = region_name,
    aws_access_key_id = aws_access_key_id,
    aws_secret_access_key = aws_secret_access_key,
    )


# Initialize the Bedrock embedding model.
# `amazon.titan-text-express-v1` is the specific model used to convert text into numerical vectors.
bedrock_embedding = BedrockEmbeddings(model_id="amazon.titan-text-express-v1", client= bedrock)


# Defines a function to load and process documents.
def get_documents():
    # Loads all PDF files from the "Data" directory.
    loader = PyPDFDirectoryLoader("Data")
    documents = loader.load()
    # Initializes the text splitter with a specific chunk size and overlap.
    # This is crucial for breaking down large documents into manageable pieces for the LLM.
    text_spliter = RecursiveCharacterTextSplitter(
                                        chunk_size=1000, 
                                        chunk_overlap=500)
    # Splits the loaded documents into chunks.
    docs = text_spliter.split_documents(documents)
    return docs


# Defines a function to create a vector store.
def get_vector_store(docs):
   # Creates a Chroma vector store from the document chunks and embeddings.
   # `persist_directory` saves the vector store to disk, so it doesn't have to be recreated every time.
    vectordb = Chroma.from_documents(docs, embedding=bedrock_embedding, persist_directory='./db')
    # Persists the vector store to the specified directory.
    vectordb.persist()


# Defines a function to initialize the LLM.
def get_llm():
    # Initializes the LLM using the Mistral 7B model from Bedrock.
    llm = Bedrock(model_id = "mistral.mistral-7b-instruct-v0:2", client = bedrock)
    return llm


# Creates a PromptTemplate object using the defined template and input variables.
PROMPT = PromptTemplate(
    template=prompt_template, input_variables=["context", "question"]
)


# Defines the core function for generating a response using the RAG chain.
def get_llm_response(llm, vectorstore, query):
    # Sets up the RetrievalQA chain.
    qa = RetrievalQA.from_chain_type(
        llm = llm,
        # `stuff` chain type combines all retrieved documents into a single prompt.
        chain_type = "stuff",
        # Configures the retriever to use similarity search on the vector store.
        # It retrieves the top 3 most relevant documents (k=3).
        retriever= vectorstore.as_retriever(
        search_type="similarity", search_kwargs={"k": 3}),
        # Ensures that the source documents used to generate the answer are returned.
        return_source_documents = True,
        # Passes the custom prompt template to the chain.
        chain_type_kwargs={"prompt": PROMPT})

    # Executes the query through the RAG chain.
    response = qa({"query": query})
    # Returns only the generated text result.
    return response['result']


# Defines the main Streamlit application function.
def main():
    # Sets the basic configuration for the Streamlit page.
    st.set_page_config("RAG")
    st.header("End to end RAG using Bedrock")

    # Creates a text input field for the user to enter their question.
    user_question = st.text_input("Ask a question from the PDF file")

    # Creates a sidebar for controls.
    with st.sidebar:
        st.title("Update & create vectore store")

        # Button to trigger the document processing and vector store creation.
        if st.button("Store Vector"):
            with st.spinner("Processing.."):
                docs = get_documents()
                get_vector_store(docs)
                st.success("Done")

        # Button to send the user's question and get a response.
        if st.button("Send"):
            with st.spinner("Processing.."):
               # Loads the pre-existing Chroma vector store from the disk.
               vectordb = Chroma(persist_directory="db",embedding_function=bedrock_embedding)
               llm = get_llm()
               # Displays the response generated by the RAG chain.
               st.write(get_llm_response(llm,vectordb,  user_question))

# Entry point of the script.
if __name__ == "__main__":
    main()
```