Amazon Bedrock/RAG Project/requirements.txt
langchain
langchain-community
pypdf
streamlit
chromadb==0.4.4
boto3
python-dotenv
